Fixed duplicate domain processing, and moved extra functions out of *_parse.py
